refactor(parser): move all remaining AST nodes
This had to happen in a single commit to ensure that circular
dependencies were correclty resolved.

## TODO:
- Increase converage
- Fix poor quality code Issues
- Simpligy certain parts of the AST nodes
