+++ fortls/parsers/internal/include.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from .scope import Scope
+
+
+class Include(Scope):
+    def get_desc(self):
+        return "INCLUDE"
