+++ fortls/parsers/internal/program.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from .module import Module
+
+
+class Program(Module):
+    def get_desc(self):
+        return "PROGRAM"
