+++ setup.cfg
@@ -48,7 +48,7 @@ install_requires =
     typing-extensions; python_version < "3.8"
 
 [options.package_data]
-fortls = *.json
+fortls = parsers/internal/*.json
 
 [options.entry_points]
 console_scripts =
